requirements.txt
requests
keyboard
Pillow
tkinter